Namespace commands with `zendesk` or `zd`
We run multiple :hubot: packages and have found it eaiser to look up the commands and run them in a non-conflicting way if we add a keyword describing the service to the front of it. So for example:

`hubot tickets`

becomes:

`hubot zendesk tickets`

With the use of an alias, such as `!`, it shortens it up nicer to where the command is more directly relatable:

`!zendesk tickets`

At any rate, this PR is a suggestion. It obviously breaks userspace, but it will not effect anyoen that used the `npm install hubot-zendesk --save` method, as that will lock it down to a particular version.

**Other changes:**

- Wired up the test suite so you could hook it into Travis easily (`npm test`)
- Removed the redundant note about adding it to package.json. Running the `--save` method described above already does that for you.
- Adjusted some code formatting in the README